requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
scikit-learn
matplotlib
